Stop trying to make fetch happen
AAD will never be useful for storing
data so give up and switch to Cosmos

Create a key-value store abstraction
on top of Cosmos that will always
have a valid document for any key
(because it will generate one if it
does not exist).

Back this up with a memory cache and
use it for profile data.